deleting toast and change base url

+++ app/src/main/java/com/capstoneC23PS274/segar/data/remote/response/CheckResponse.kt
@@ -17,7 +17,7 @@ data class CheckResponse(
 data class CheckResult(
 
 	@field:SerializedName("score")
-	val score: Int? = null,
+	val score: Double? = null,
 
 	@field:SerializedName("image")
 	val image: String? = null,
